linked issue: https://github.com/shopspring/decimal/issues/81

before change:
```
goos: darwin
goarch: amd64
pkg: github.com/ngalayko/decimal
Benchmark_decimal_Decimal_Add_different_precision-4      2000000               822 ns/op             511 B/op         12 allocs/op
Benchmark_decimal_Decimal_Sub_different_precision-4      2000000               865 ns/op             496 B/op         12 allocs/op
Benchmark_math_big_Float_Sub_different_precision-4      100000000               10.9 ns/op             0 B/op          0 allocs/op
Benchmark_math_big_Float_Add_different_precision-4      20000000                79.7 ns/op             2 B/op          0 allocs/op
Benchmark_decimal_Decimal_Add_same_precision-4           2000000               642 ns/op             430 B/op         10 allocs/op
Benchmark_decimal_Decimal_Sub_same_precision-4           2000000               659 ns/op             427 B/op         10 allocs/op
Benchmark_math_big_Float_Add_same_precision-4           10000000               172 ns/op              48 B/op          1 allocs/op
Benchmark_math_big_Float_Sub_same_precision-4           10000000               179 ns/op              48 B/op          1 allocs/op
```

after change:
```
goos: darwin
goarch: amd64
pkg: github.com/ngalayko/decimal
Benchmark_decimal_Decimal_Add_different_precision-4      2000000               843 ns/op             511 B/op         12 allocs/op
Benchmark_decimal_Decimal_Sub_different_precision-4      2000000               999 ns/op             496 B/op         12 allocs/op
Benchmark_math_big_Float_Sub_different_precision-4      200000000                7.91 ns/op            0 B/op          0 allocs/op
Benchmark_math_big_Float_Add_different_precision-4      20000000                81.9 ns/op             0 B/op          0 allocs/op
Benchmark_decimal_Decimal_Add_same_precision-4          10000000               131 ns/op              80 B/op          2 allocs/op
Benchmark_decimal_Decimal_Sub_same_precision-4          10000000               137 ns/op              85 B/op          2 allocs/op
Benchmark_math_big_Float_Add_same_precision-4           10000000               177 ns/op              48 B/op          1 allocs/op
Benchmark_math_big_Float_Sub_same_precision-4           10000000               184 ns/op              48 B/op          1 allocs/op
```